refactor: relocate AuthorResolver to the application layer (PR2a)
Pure move, no behaviour change: AuthorResolver goes from BookTracker.Web/Services to BookTracker.Application/Authors so the upcoming Works command handlers can resolve authors without reaching up into Web. Only depends on Data types, so the move is clean.

Callers (BookDetail/WorkEdit/BookAdd/BulkAdd/Wishlist VMs + AuthorResolverTests) get the new using; BulkAdd.razor only names it in a comment, so it's untouched. Full suite: 558 passed, 1 skipped — unchanged from before the move.

Next (PR2b): AssignAuthors' logic moves onto the Work aggregate as AssignAuthorship, and the Works handlers build on this.
