+++ BookTracker.Tests/Services/AuthorResolverTests.cs
@@ -1,5 +1,5 @@
+using BookTracker.Application.Authors;
 using BookTracker.Data.Models;
-using BookTracker.Web.Services;
 
 namespace BookTracker.Tests.Services;
