+++ BookTracker.Web/ViewModels/BookAddViewModel.cs
@@ -1,3 +1,4 @@
+using BookTracker.Application.Authors;
 using BookTracker.Data;
 using BookTracker.Data.Models;
 using BookTracker.Web.Services;
